collatz improve2

+++ collatz.py
@@ -6,10 +6,10 @@ list=[x]
 
 while x!=1:
     if x%2==0:        
-        x = x/2
+        x = int(x/2)
         list.append(x)
     else: 
-        x = x*3+1
+        x = int(x*3+1)
         list.append(x)
 
 print(list)
